+++ src/leapmotor_api/const.py
@@ -55,5 +55,7 @@ REMOTE_CTL_REARVIEW_MIRROR_HEAT = "rearview_mirror_heat"
 REMOTE_CTL_REARVIEW_MIRROR_HEAT_ON = "rearview_mirror_heat_on"
 REMOTE_CTL_REARVIEW_MIRROR_HEAT_OFF = "rearview_mirror_heat_off"
 REMOTE_CTL_SPEED_LIMIT = "set_speed_limit"
+REMOTE_CTL_SEAT_HEAT = "seat_heat"
+REMOTE_CTL_SEAT_VENTILATION = "seat_ventilation"
 
 DEFAULT_TIMEOUT = 30